포스트 콘텐츠 렌더링하는 컴포넌트 추출
