+++ src/components/pages/post/PostDetailPage.tsx
@@ -1,5 +1,5 @@
 import { deleteDoc, doc } from 'firebase/firestore';
-import { AlertCircle, Edit, Trash2, Lock } from 'lucide-react';
+import { Edit, Trash2 } from 'lucide-react';
 import { useNavigate, useParams, Link } from 'react-router-dom';
 import { Button } from '@/components/ui/button';
 import { Skeleton } from '@/components/ui/skeleton';
@@ -11,31 +11,8 @@ import { fetchPost } from '../../../utils/postUtils';
 import Comments from '../comment/Comments';
 import { PostBackButton } from './PostBackButton';
 import { PostAdjacentButtons } from './PostAdjacentButtons';
-import { sanitizePostContent } from '@/utils/contentUtils';
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
-import * as Sentry from '@sentry/react';
-import { PostVisibility } from '@/types/Posts';
 import { formatDateToKorean } from '@/utils/dateUtils';
-
-const deletePost = async (boardId: string, id: string): Promise<void> => {
-  await deleteDoc(doc(firestore, `boards/${boardId}/posts`, id));
-};
-
-const handleDelete = async (
-  postId: string,
-  boardId: string,
-  navigate: (path: string) => void,
-): Promise<void> => {
-  const confirmDelete = window.confirm('정말로 이 게시물을 삭제하시겠습니까?');
-  if (!confirmDelete) return;
-
-  try {
-    await deletePost(boardId, postId);
-    navigate(`/board/${boardId}`);
-  } catch (error) {
-    console.error('게시물 삭제 오류:', error);
-  }
-};
+import { PostContent } from './PostContent';
 
 export default function PostDetailPage() {
   const { postId, boardId } = useParams<{ postId: string; boardId: string }>();
@@ -83,70 +60,6 @@ export default function PostDetailPage() {
 
   const isAuthor = currentUser?.uid === post.authorId;
 
-  // Private 포스트는 작성자만 볼 수 있음
-  if (post.visibility === PostVisibility.PRIVATE && !isAuthor) {
-    return (
-      <div className='mx-auto max-w-4xl px-6 sm:px-8 lg:px-12 py-8 text-center'>
-        <Alert className="mb-6 border-yellow-500 bg-yellow-50">
-          <Lock className="size-4 text-yellow-800" />
-          <AlertTitle className="text-yellow-800">비공개 게시글</AlertTitle>
-          <AlertDescription className="text-yellow-700">
-            이 게시글은 작성자만 볼 수 있습니다.
-          </AlertDescription>
-        </Alert>
-        {boardId && <PostBackButton boardId={boardId} />}
-      </div>
-    );
-  }
-
-  const sanitizedContent = sanitizePostContent(post.content);
-  const renderContent = () => {
-    if (!post?.content) {
-      return <p>내용이 없습니다.</p>;
-    }
-
-    try {
-      return (
-        <div
-          dangerouslySetInnerHTML={{ __html: sanitizedContent }}
-          className="prose prose-lg prose-slate dark:prose-invert max-w-none mt-6
-            prose-h1:text-3xl prose-h1:font-semibold 
-            prose-h2:text-2xl prose-h2:font-semibold
-            prose-p:my-4
-            prose-ul:my-4
-            prose-ol:my-4
-            "
-        />
-      );
-    } catch (error) {
-      const err = error instanceof Error ? error : new Error('알 수 없는 렌더링 오류가 발생했습니다.');
-      
-      Sentry.captureException(err, {
-        extra: {
-          postId,
-          boardId,
-        },
-      });
-      
-      return (
-        <Alert variant="destructive" className="my-4">
-          <AlertCircle className="size-4" />
-          <AlertTitle>렌더링 오류</AlertTitle>
-          <AlertDescription className="mt-2">
-            <p>콘텐츠를 화면에 표시하는 중에 문제가 발생했습니다:</p>
-            <p className="mt-1 text-sm font-mono bg-red-50 p-2 rounded">
-              {err.message}
-            </p>
-            <p className="mt-2 text-sm">
-              페이지를 새로고침하거나 나중에 다시 시도해주세요.
-              문제가 계속되면 관리자에게 문의해주세요.
-            </p>
-          </AlertDescription>
-        </Alert>
-      );
-    }
-  };
-
   return (
     <div className='mx-auto max-w-4xl px-6 sm:px-8 lg:px-12 py-8'>
       {boardId && <PostBackButton boardId={boardId} className='mb-6' />}
@@ -154,11 +67,6 @@ export default function PostDetailPage() {
         <header className='space-y-4'>
           <div className="flex items-center gap-2">
             <h1 className='text-4xl font-bold leading-tight tracking-tight text-gray-900 dark:text-gray-100 sm:text-5xl mb-4'>{post.title}</h1>
-            {post.visibility === PostVisibility.PRIVATE && (
-              <span className="inline-flex items-center rounded-md bg-yellow-50 px-2 py-1 text-xs font-medium text-yellow-800 ring-1 ring-inset ring-yellow-600/20">
-                <Lock className="size-3 mr-1" /> 이 글은 작성자만 볼 수 있어요
-              </span>
-            )}
           </div>
           <div className='flex items-center justify-between text-sm text-gray-500 dark:text-gray-400'>
             <p>
@@ -184,7 +92,7 @@ export default function PostDetailPage() {
             )}
           </div>
         </header>
-        {renderContent()}
+        <PostContent post={post} isAuthor={isAuthor} />
       </article>
       <div className='mt-12 border-t border-gray-200'></div>
       {boardId && postId && <PostAdjacentButtons boardId={boardId} postId={postId} />}
@@ -194,3 +102,23 @@ export default function PostDetailPage() {
     </div>
   );
 }
+
+const deletePost = async (boardId: string, id: string): Promise<void> => {
+  await deleteDoc(doc(firestore, `boards/${boardId}/posts`, id));
+};
+
+const handleDelete = async (
+  postId: string,
+  boardId: string,
+  navigate: (path: string) => void,
+): Promise<void> => {
+  const confirmDelete = window.confirm('정말로 이 게시물을 삭제하시겠습니까?');
+  if (!confirmDelete) return;
+
+  try {
+    await deletePost(boardId, postId);
+    navigate(`/board/${boardId}`);
+  } catch (error) {
+    console.error('게시물 삭제 오류:', error);
+  }
+};
